chore: ignore TS build info; untrack tsconfig.tsbuildinfo
