+++ frontend/src/__tests__/components/ExportImportControls.test.tsx
@@ -1,69 +1,30 @@
-// __tests__/ExportImportControls.test.tsx
+// __tests__/ExportImportControls.test.tsx (updated)
 import { render, screen, fireEvent } from "@testing-library/react";
 import ExportImportControls from "@components/ExportImportControls";
-import { exportPortfolio } from "@utils/exportPortfolio";
-import { getFormattedDate } from "@utils/formatDate";
-
 import { vi } from "vitest";
 
-// Mock dependencies
-vi.mock("@utils/exportPortfolio", () => ({
-  exportPortfolio: vi.fn(() => "mocked,data,output"),
-}));
-
-vi.mock("@utils/formatDate", () => ({
-  getFormattedDate: vi.fn(() => "2099-12-31"),
-}));
-
 describe("ExportImportControls", () => {
-  const portfolio = [
-    { asset: "btc", quantity: 1 },
-    { asset: "eth", quantity: 2 },
-  ];
-  const prices = {
-    btc: 30000,
-    eth: 2000,
-  };
+  it("calls onExport with the selected format and onImport when clicked", () => {
+    const onExport = vi.fn();
+    const onImport = vi.fn();
 
-  beforeEach(() => {
-    // Mock URL.createObjectURL and revokeObjectURL
-    global.URL.createObjectURL = vi.fn(() => "blob:mock-url");
-    global.URL.revokeObjectURL = vi.fn();
+    render(<ExportImportControls onExport={onExport} onImport={onImport} />);
 
-    // Spy on <a>.click()
-    const originalCreateElement = document.createElement;
+    // Default format is CSV per component state
+    fireEvent.click(screen.getByTestId("export-button"));
+    expect(onExport).toHaveBeenCalledWith("csv");
 
-    vi.spyOn(document, "createElement").mockImplementation(
-      (tagName: string) => {
-        if (tagName === "a") {
-          return {
-            set href(value: string) {},
-            set download(value: string) {},
-            click: vi.fn(),
-          } as unknown as HTMLAnchorElement;
-        }
-        return originalCreateElement.call(document, tagName);
-      }
-    );
-  });
+    // Change to JSON and export again
+    const select = screen.getByLabelText(/select file format/i);
+    fireEvent.change(select, { target: { value: "json" } });
+    fireEvent.click(screen.getByTestId("export-button"));
+    expect(onExport).toHaveBeenCalledWith("json");
 
-  afterEach(() => {
-    vi.restoreAllMocks();
-  });
-
-  it("triggers export with correct blob and filename", () => {
-    render(<ExportImportControls portfolio={portfolio} prices={prices} />);
-
-    const button = screen.getByRole("button", {
-      name: /download portfolio as csv/i,
-    });
-
-    fireEvent.click(button);
-
-    expect(exportPortfolio).toHaveBeenCalledWith(portfolio, prices, "csv");
-    expect(getFormattedDate).toHaveBeenCalled();
-
-    expect(URL.createObjectURL).toHaveBeenCalledWith(expect.any(Blob));
-    expect(URL.revokeObjectURL).toHaveBeenCalled();
+    const file = new File([JSON.stringify([{ asset: "btc", quantity: 1 }])], "portfolio.json", { type: "application/json" });
+    const input = screen.getByTestId("import-file-input") as HTMLInputElement;
+    Object.defineProperty(input, "files", { value: [file] });
+    fireEvent.change(input);
+    fireEvent.click(screen.getByTestId("import-button"));
+    expect(onImport).toHaveBeenCalled();
   });
 });
