--- frontend/tsconfig.tsbuildinfo
@@ -1,1 +1,0 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/setuptests.ts","./src/vite-env.d.ts","./src/__context/coincontext.tsx","./src/__context/coinprovider.tsx","./src/__context/usecoincontext.ts","./src/__tests__/app-pages/protfolio-page.test.tsx","./src/__tests__/components/addassetmodal.test.tsx","./src/__tests__/components/assetlist.test.tsx","./src/__tests__/components/assetselector.test.tsx","./src/__tests__/hooks/useaddassetform.test.ts","./src/__tests__/hooks/useassetlist.test.ts","./src/__tests__/hooks/usecoinlist.test.ts","./src/__tests__/hooks/usecoinsearch.test.ts","./src/__tests__/hooks/usepolling.test.ts","./src/__tests__/hooks/useportfoliostate.test.ts","./src/__tests__/hooks/usepricemap.test.ts","./src/__tests__/hooks/useuistate.test.ts","./src/__tests__/services/coinservice.test.ts","./src/__tests__/utils/validateasset.test.ts","./src/components/addassetmodal.tsx","./src/components/assetselector.tsx","./src/components/errorbanner.tsx","./src/components/inlineerrorbadge.tsx","./src/components/loadingspinner.tsx","./src/components/assetlist/index.tsx","./src/components/assetrow/assetrow.test.tsx","./src/components/assetrow/index.tsx","./src/components/deleteconfirmationmodal/deleteconfirmationmoda.test.tsx","./src/components/deleteconfirmationmodal/index.tsx","./src/components/exportimportcontrols/exportimportcontrols.test.tsx","./src/components/exportimportcontrols/index.tsx","./src/components/filtersortcontrols/filtersortcontrols.test.tsx","./src/components/filtersortcontrols/index.tsx","./src/components/portfolioheader/portfolioheader.test.tsx","./src/components/portfolioheader/index.tsx","./src/components/__mocks__/mockcoinprovider.tsx","./src/components/__mocks__/fixtures.ts","./src/components/__stories__/ assetrow.stories.tsx","./src/components/__stories__/addassetmodal.stories.tsx","./src/components/__stories__/deleteconfirmationmoda.stories.tsx","./src/e2e/fixtures.ts","./src/e2e/test-setup.ts","./src/e2e/mocks/mockcoingecko.ts","./src/e2e/pom-pages/add-asset-modal.pom.ts","./src/e2e/pom-pages/delete-confirmation-modal.pom.ts","./src/e2e/pom-pages/portfolio.pom.ts","./src/e2e/specs/a11y-contrast.spec.ts","./src/e2e/specs/addasset.spec.ts","./src/e2e/specs/calculate-total-value.spec.ts","./src/e2e/specs/delete-asset.spec.ts","./src/e2e/specs/portfolio-layout.spec.ts","./src/e2e/specs/portfolio-value.spec.ts","./src/e2e/utils/aa11y-check.ts","./src/hooks/useaddassetform.ts","./src/hooks/usecoinlist.ts","./src/hooks/usecoinsearch.ts","./src/hooks/usepolling.ts","./src/hooks/useportfoliostate.ts","./src/hooks/usepricemap.ts","./src/hooks/useuistate.ts","./src/pages/portfoliopage.tsx","./src/services/coinservice.ts","./src/utils/index.ts","./src/utils/validateasset.ts"],"errors":true,"version":"5.8.3"}
